FLGO-136 refactor api

+++ app/shared/config.go
@@ -2,9 +2,9 @@ package shared
 
 type AppConfig struct {
 	ImagesBaseURL         string
-	BlogsInPage           uint16
-	BlogTopicsInPage      uint16
-	ForumTopicsInPage     uint32
-	ForumMessagesInPage   uint32
+	BlogsInPage           uint64
+	BlogTopicsInPage      uint64
+	ForumTopicsInPage     uint64
+	ForumMessagesInPage   uint64
 	DefaultAccessToForums []uint16
 }
